Added new tools, and reworkd clang-tidy usage
Added two new tools, include-what-you-use and cppcheck, both of which
are disabled by default. They can be enabled via two new variable
options.

clang-tidy is now enabled by default, and runs with the compiler instead
of a separate run, as the separate run was rife with issues to setup,
sometimes causing infinite loops in CMake configuration stage. It can
be disabled via a new variable.