Made local catch header search more robust
When searching for the catch header locally, the script now searches
the catch/ subfolder, where most installations locate the headers in
filesystems rather than being a top-level header.

As a consequence, modified the script to also take the option to force
the cloning of a catch header if local headers prove difficult.