Modified tests macro and Catch header inclusion
Now searches for the catch2 directory as well as the original catch
directory. Also searches for Catch headers on initial configuration
instead of after the BUILD_TESTS option is enabled. This means one less
configuration step required, as well as showing if one is found before
attempting to enable tests.

Renamed the cloning target from 'catch2' to 'git_catch', for clarity.