Reworked clang-tidy tool configuration
When using a configuration tool such as ccmake, the use would have to
perform two configurations before a generation could be performed. This
was due to the setting CMAKE_CXX_CLANG_TIDY being set only if the
CLANG_TIDY variable was set, which would normally be after the first
config found the option.

Moving the setting of the variable to if/when clang-tidy program itself
was found allows for full project generation after only the first config
step.