Updated example/test inclusion checking
When testing for adding example or tests, CMake would only check for the
existence of the subfolder, rather than if a CMakeLists.txt file existed
within that folder. This often lead to issues where the empty directory
would cause CMake to error out.

This fixes that by checking for the CMake file directly, and only then
adding that subdirectory.