Reworked code coverage for flexibility
Code coverage has previously worked on the basis of being enabled on an
all-or-nothing basis, for all targets or none.

This was problematic for cases where coverage information was wanted
only for particular targets, or for libraries but not the test programs
testing them.

To alleviate this, the macros have been revised. to allow for either
blanket coverage or per-target coverage settings. As well, an improved
usage guide has been added to the file.