Added recursive guard for code coverage
If the cmake file is used in a recursive manner, such as in a
subdirectory or submodule, the code coverage would fail out with an
error due to an attempt to re-add the same ccov-* targets. Adding a
guard prevents this from occurring by preventing re-initialization of
the same preparatory targets.