Converged both formatting tools to one file
The 'formatting.cmake' file now has both the older clang-format and the
new cmake-format. This will hopefully also house other formatters as
they are added. The README has been updated accordingly.

The formatters have been updated, so that both accept file paths that
are either absolute or relative to the current source dir. Also, all
files to be formatted are set as individual targets, that can still be
grouped.