More changes/fixes for formatting
First, reduced the number of targets for formatting. While previously
each file got it's own target, while fine in makefiles, when using with
Visual Studio or similar, this exploded the number of targets to a
ridiculous degree.

Thus, now only the specific named targets, and all the files are
collapsed into them. As well, removed the silliness that was
surreptitously adding prefixes to format target names. Now, only the
given target name is used, and errors out if the name is already in use.