Fixed catch shared library issues
The catch library, when in a scope that used shared object defaulting to
'hidden' would fail due to the symbols being, well, hidden. To fix, the
build_tests macro has been subdivided into a sub-function so that the
visibility can be reset to visible, without affecting outset scope. Also
changed non WIN32 builds of the catch pre-compiled library to always be
shared.