Code coverage shared object support much improved
As with the earlier commit that added shared object support for llvm-cov
for the 'all'-type targets, this one adds that similar support to
individual targets.

This involves the new option of 'OBJECTS' on the target_code_coverage
calls that explicitly state which shared objects to be instrumented by
which executable calls.