Rework test.cmake
Changed build_tests function to prepare_catch.

The new function is just the old  build_tests_internal function, but
without the enable_testing and add_subdirectory calls. This allows for
it to be used more flexibly as it has less baked-in behaviour, and now
only encompasses the pulling in and preparation of the Catch framework.