Added formatting checks to CI/CD
For both clang-format and cmake-format, both providing examples as
to how to do it within a CI/CD and to actually keep the code here
in good shape.

Using the `git diff --exit-code` rather than the built-in check
functionality allows for both checking for changes and an exit code
but also prints out the changes required to fix the issues, useful
CI output for users.