Changed pipeline jobs to reflect proper success
THe sanitizer jobs which were to test failures did so by actually
failing the jobs successfully. However, this leads to the issue of the
CI pipeline always showing as 'passed with warnings'.

THe jobs that are supposed to fail now invert the final test results if
they fail successfully. Tests that are supposed to pass do so.

This now leads to a CI pipeline that more accurately reflects the state
of the repository and tests.