Marked several variables as 'advanced'
Many of the variables generated by CMake when searching for executables
for items such as clang-format or such would remain and pollute the GUI
or ccmake view.

By marking such variables as 'advanced', these variables are hidden by
default, but can be easily displayed by entering advanced mode.

In these cases it's fairly safe, since error messages occur if the items
in question cannot be found.