Fixed shared code not showing on ccov-all reports
When collating the data from the individual runs, the only binary
objects listed were for the test executables themselves.

This was fine for statically linked code, but shared code is external to
these, and wouldn't show up on the final reports.

Thus, adding the executable's shared object links to the all target's
`binaries.list` file fixed this.