Fixed ccov-all targets on Windows
When running ccov-run-* targets in parrallel on
Windows, multiple processes would attempt to write to the profile data
files, causing contention and failure. This job has been split in two
and the call to write to file has been serialized through a specified
job pool 'ccov_serial_pool'.

For the ccov-all targets, added a Windows-specific set that utilizes
Powershell commands to work.