Add AFL++ fuzzing instrumentation capability
For AFL itself, when enabled via the AFL variable will search for the
base afl-cc/afl-c++ compiler wrappers, and if found use them. Then any
options specified at configuration time such as the speicifc mode or
options must be specified.

Because of the nature of having to change the compiler, this file must
be included at a very early stage before the compilers are detected and
verified.

Since it supports sanitizers, the sanitizers file has been updated to
add the appropriate flags when AFL is enabled.