Rework of tools.cmake items
Previously, the turning on/off clang-tidy, include-what-you-use and
cppcheck was handled by a combination of an option and the use of macros
for specific details. This has meant only straightforward usage was
accomodated.

Rather, there is a desire to be more flexible, to allow for difference
code to be compiled with different options, or to turn on/off tools
altogether for specific scopes.

The options of `CLANG_TIDY`, `IWYU` and `CPPCHECK` have been removed.
Instead, it is up to the including project on when/how to enable these
tools. The ability to 'reset' the tools, to disable them for certain
scopes has been accomodated via the use of `reset_*` macros.