Modernize sanitizer scripts
The previous sanitizer support was implemented in a very rigid manner,
with the compile options available for each sanitizer type been preset,
and enabling of each sanitizer through a preset variable.

This change makes it so that setting of compile options can be set on a
per-sanitizer-type basis and then added to builds via functions. Use via
the legacy `USE_SANITIZER` is available, but deprecated.

This is also flexible enough to allow overriding of the compile options,
via variables either in-script or passed in via the command-line.

Results of checks for compile compatability are cached, reducing time
on subsequent builds, provided the associated compile options are not
updated.

Checks for sanitizer availability/compatability are no longer platform
dependent. It will check on all platforms.