Improved code-coverage parallelization
An issue with LLVM-based code coverage has been the fact that, if run in
parallel, multiple targets could be be writing to the binaries.list
and/or profraw.list files simultaneously, leading to undefined results,
or on Windows the command failing due to the file being used by another
process entirely.

To avoid this, each target will now write to separate target-specific
files in the coverage data directories. Then when the data is actually
required, a few `cat` commands will run collecting all the items into
single files for further use.