Rework ccov-run* coverage targets
Previously, any time a coverage target was called, the executables would
always be run. Recently, this was changes so that the executables were
run to generate the coverage data and not re-run if the underlying
executable had not been changed.

This commit change it once again to have both. At a minimum the
executables will be run to generate the data, and will not be re-run
unless either the dependencies have been changed/updated OR the
ccov-run* targets are invoked.